Added new check to resamplers.py for zero-norm cov.
I noted in testing that sometimes an updater would get very
close to the true model, but would fail with a ResamplerError
due to the estimated norm of the error in sqrtm(cov) being nan.
This can happen either if the error term is awful, or if the
norm of cov is zero; the error is intended for the former case,
not the latter. To compensate, the LW resampler as of this commit
now checks for zero-norm cov and if detected, warns, then adds
a small artificial covariance to compensate.